Autenticação com token

+++ app/Http/Controllers/ResidenceImageController.php
@@ -15,47 +15,83 @@ class ResidenceImageController extends Controller
      */
     public function index()
     {
-        $all = ResidenceImage::with('residence')->get();
-        return response()->json($all);
+        $var_token = $request->get('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = ResidenceImage::with('residence')->get();
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     public function high($idResidenceImage)
     {
-        $residenceImage = ResidenceImage::with('residence')
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+             $residenceImage = ResidenceImage::with('residence')
             ->where('idResidenceImage', '=', $idResidenceImage)
             ->orderBy('orderImage','desc')
             ->first(); 
-        $path = public_path()."/img/residenceImage/high/".$residenceImage['path'];
-        
-        $data = base64_encode(file_get_contents($path));
-        $src = 'data: '.mime_content_type($path).';base64,'.$data;
-        echo '<img src="'.$src.'">';
+            $path = public_path()."/img/residenceImage/high/".$residenceImage['path'];
+            
+            $data = base64_encode(file_get_contents($path));
+            $src = 'data: '.mime_content_type($path).';base64,'.$data;
+            echo '<img src="'.$src.'">';
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
      public function medium($idResidenceImage)
     {
-        $residenceImage = ResidenceImage::with('residence')
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+             $residenceImage = ResidenceImage::with('residence')
             ->where('idResidenceImage', '=', $idResidenceImage)
             ->orderBy('orderImage','desc')
             ->first(); 
-        $path = public_path()."/img/residenceImage/medium/".$residenceImage['path'];
-        
-        $data = base64_encode(file_get_contents($path));
-        $src = 'data: '.mime_content_type($path).';base64,'.$data;
-        echo '<img src="'.$src.'">';
+            $path = public_path()."/img/residenceImage/medium/".$residenceImage['path'];
+            
+            $data = base64_encode(file_get_contents($path));
+            $src = 'data: '.mime_content_type($path).';base64,'.$data;
+            echo '<img src="'.$src.'">';
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
      public function low($idResidenceImage)
     {
-        $residenceImage = ResidenceImage::with('residence')
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+             $residenceImage = ResidenceImage::with('residence')
             ->where('idResidenceImage', '=', $idResidenceImage)
             ->orderBy('orderImage','desc')
             ->first(); 
-        $path = public_path()."/img/residenceImage/low/".$residenceImage['path'];
-        
-        $data = base64_encode(file_get_contents($path));
-        $src = 'data: '.mime_content_type($path).';base64,'.$data;
-        echo '<img src="'.$src.'">';
+            $path = public_path()."/img/residenceImage/low/".$residenceImage['path'];
+            
+            $data = base64_encode(file_get_contents($path));
+            $src = 'data: '.mime_content_type($path).';base64,'.$data;
+            echo '<img src="'.$src.'">';
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -65,10 +101,19 @@ class ResidenceImageController extends Controller
      */
     public function residence($idResidence)
     {
-        $residence = ResidenceImage::with('residence')
-        ->where('idResidence', '=', $idResidence)
-        ->get();
-        return response()->json($residence);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residence = ResidenceImage::with('residence')
+            ->where('idResidence', '=', $idResidence)
+            ->get();
+            return response()->json($residence);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -78,11 +123,19 @@ class ResidenceImageController extends Controller
      */
     public function topAllResidences()
     {
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
 
-        $residence = ResidenceImage::with('residence')
-        ->orderBy('orderImage','desc')
-        ->groupBy('idResidence')->get();     
-        return response()->json($residence);
+        if($user_token != null) {
+            $residence = ResidenceImage::with('residence')
+            ->orderBy('orderImage','desc')
+            ->groupBy('idResidence')->get();     
+            return response()->json($residence);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -92,14 +145,22 @@ class ResidenceImageController extends Controller
      */
     public function topAllResidencesLimit($id,$qtReg)
     {
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
 
-        $residence = ResidenceImage::with('residence')
-        ->orderBy('idResidence','asc')
-        ->groupBy('idResidence')
-        ->limit($qtReg)
-        ->offset($id-1)
-        ->get();     
-        return response()->json($residence);
+        if($user_token != null) {
+            $residence = ResidenceImage::with('residence')
+            ->orderBy('idResidence','asc')
+            ->groupBy('idResidence')
+            ->limit($qtReg)
+            ->offset($id-1)
+            ->get();     
+            return response()->json($residence);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -109,11 +170,20 @@ class ResidenceImageController extends Controller
      */
     public function residenceTop($idResidence)
     {
-        $residence = ResidenceImage::with('residence')
-        ->where('idResidence', '=', $idResidence)
-        ->orderBy('orderImage', 'desc')
-        ->first();     
-        return response()->json($residence);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residence = ResidenceImage::with('residence')
+            ->where('idResidence', '=', $idResidence)
+            ->orderBy('orderImage', 'desc')
+            ->first();     
+            return response()->json($residence);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -124,9 +194,50 @@ class ResidenceImageController extends Controller
      */
     public function store(Request $request)
     {
-        $residenceImage = $request->all();
-        $residenceImage = ResidenceImage::create($residenceImage);
-        return response()->json($residenceImage);
+       $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residenceImageReceive = $request->get('jsonObject');
+            //retorna array, se tirar o true do json_decode vai retornar um objeto e vai dar erro em tudo.
+            $residenceImageReceive = json_decode($residenceImageReceive,true);
+            
+            $residence = $residenceImageReceive['residence'];
+
+            //Adiciona as imagens do usuario.
+               $residenceImage = $residenceImageReceive;
+               //foreach($userImages as $userImage){
+                   $file = $residenceImage['path'];
+                  // $size = pathinfo($file);
+                  // dd($size);
+                   $file = base64_decode($file);
+
+                    $name_image = "userImage_".time().".png";
+                    $path_high = public_path()."/img/residenceImage/high/".$name_image;
+                    $path_medium = public_path()."/img/residenceImage/medium/".$name_image;
+                    $path_low = public_path()."/img/residenceImage/low/".$name_image;
+
+                    Image::make($file)->save($path_high);
+                    $size = getimagesize(   public_path()."/img/residenceImage/high/".$name_image);
+                    Image::make($file)->resize($size[0]/3, $size[1]/3)->save($path_medium);
+                    Image::make($file)->resize($size[0]/5, $size[1]/5)->save($path_low);
+
+                    $residenceImageReceive = ResidenceImage::create([
+                        'path' => $name_image,
+                       // 'resource' => $userImage['resource'],
+                        'orderImage' => $userImage['orderImage'],
+                        'idUser' => $user['idUser'],
+                        'tries' => 0,
+                        'active' => true,
+                        ]);   
+               // }
+
+            return response()->json($residenceImageReceive);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -137,8 +248,17 @@ class ResidenceImageController extends Controller
      */
     public function show($id)
     {
-        $all = ResidenceImage::with('residence')->findOrFail($id);
-        return response()->json($all);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = ResidenceImage::with('residence')->findOrFail($id);
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -150,10 +270,19 @@ class ResidenceImageController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $residenceImage = ResidenceImage::findOrFail($id);
-        $input = $request->all();
-        $residenceImage->fill($input)->save();
-        return response()->json($residenceImage);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residenceImage = ResidenceImage::findOrFail($id);
+            $input = $request->all();
+            $residenceImage->fill($input)->save();
+            return response()->json($residenceImage);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -164,7 +293,16 @@ class ResidenceImageController extends Controller
      */
     public function destroy($id)
     {
-        $residenceImage = ResidenceImage::findOrFail($id);
-        $residenceImage->delete();
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residenceImage = ResidenceImage::findOrFail($id);
+            $residenceImage->delete();
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 }
