+++ app/Http/Controllers/UserImageController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests;
+use App\User;
 use App\UserImage;
 use Illuminate\Http\Request;
 use Intervention\Image\Facades\Image;
@@ -17,48 +18,84 @@ class UserImageController extends Controller
      */
     public function index()
     {
-        $all = UserImage::with('user')->get();
-        return response()->json($all);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = UserImage::with('user')->get();
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
-    public function high($idUserImage)
-    {
+public function high($idUserImage)
+{
+    $var_token = $_GET['api_token'];
+    $user_token = User::where('api_token',$var_token)->first();
+
+    if($user_token != null) {
         $userImage = UserImage::with('user')
-            ->where('idUserImage', '=', $idUserImage)
-            ->orderBy('orderImage','desc')
-            ->first(); 
+        ->where('idUserImage', '=', $idUserImage)
+        ->orderBy('orderImage','desc')
+        ->first(); 
         $path = public_path()."/img/userImage/high/".$userImage['path'];
-        
+
         $data = base64_encode(file_get_contents($path));
         $src = 'data: '.mime_content_type($path).';base64,'.$data;
         echo '<img src="'.$src.'">';
-    }
+    } else { return response()->json(array(
+                    'code'      =>  404,
+                    'message'   =>  'Usuário não autenticado'
+                    ), 404);
+            }
+}
 
-    public function medium($idUserImage)
-    {
+public function medium($idUserImage)
+{
+    $var_token = $_GET['api_token'];
+    $user_token = User::where('api_token',$var_token)->first();
+
+    if($user_token != null) {
         $userImage = UserImage::with('user')
-            ->where('idUserImage', '=', $idUserImage)
-            ->orderBy('orderImage','desc')
-            ->first(); 
+        ->where('idUserImage', '=', $idUserImage)
+        ->orderBy('orderImage','desc')
+        ->first(); 
         $path = public_path()."/img/userImage/medium/".$userImage['path'];
-        
-        $data = base64_encode(file_get_contents($path));
-        $src = 'data: '.mime_content_type($path).';base64,'.$data;
-        echo '<img src="'.$src.'">';
-    }
 
-    public function low($idUserImage)
-    {
-        $userImage = UserImage::with('user')
-            ->where('idUserImage', '=', $idUserImage)
-            ->orderBy('orderImage','desc')
-            ->first(); 
-        $path = public_path()."/img/userImage/low/".$userImage['path'];
-        
         $data = base64_encode(file_get_contents($path));
         $src = 'data: '.mime_content_type($path).';base64,'.$data;
         echo '<img src="'.$src.'">';
-    }
+    } else { return response()->json(array(
+                    'code'      =>  404,
+                    'message'   =>  'Usuário não autenticado'
+                    ), 404);
+            }
+}
+
+public function low($idUserImage)
+{
+    $var_token = $_GET['api_token'];
+    $user_token = User::where('api_token',$var_token)->first();
+
+    if($user_token != null) {
+        $userImage = UserImage::with('user')
+        ->where('idUserImage', '=', $idUserImage)
+        ->orderBy('orderImage','desc')
+        ->first(); 
+        $path = public_path()."/img/userImage/low/".$userImage['path'];
+
+        $data = base64_encode(file_get_contents($path));
+        $src = 'data: '.mime_content_type($path).';base64,'.$data;
+        echo '<img src="'.$src.'">';
+    } else { return response()->json(array(
+                    'code'      =>  404,
+                    'message'   =>  'Usuário não autenticado'
+                    ), 404);
+            }
+}
 
     /**
      * Store a newly created resource in storage.
@@ -68,43 +105,47 @@ class UserImageController extends Controller
      */
     public function store(Request $request)
     {
-        $userImageReceive = $request->get('jsonObject');
-        //retorna array, se tirar o true do json_decode vai retornar um objeto e vai dar erro em tudo.
-        $userImageReceive = json_decode($userImageReceive,true);
-        
-        $user = $userImageReceive['user'];
+        $var_token = $request->get('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
 
-        //Adiciona as imagens do usuario.
-           $userImage = $userImageReceive;
-           //foreach($userImages as $userImage){
-               $file = $userImage['path'];
-              // $size = pathinfo($file);
-              // dd($size);
-               $file = base64_decode($file);
+        if($user_token != null) {
+            $userImageReceive = $request->get('jsonObject');
+            //retorna array, se tirar o true do json_decode vai retornar um objeto e vai dar erro em tudo.
+            $userImageReceive = json_decode($userImageReceive,true);
+            
+            $user = $userImageReceive['user'];
 
-                $name_image = "userImage_".time().".png";
-                $path_high = public_path()."/img/userImage/high/".$name_image;
-                $path_medium = public_path()."/img/userImage/medium/".$name_image;
-                $path_low = public_path()."/img/userImage/low/".$name_image;
+            //Adiciona as imagens do usuario.
+            $userImage = $userImageReceive;
+            $file = $userImage['path'];
+            $file = base64_decode($file);
 
-                Image::make($file)->save($path_high);
-                $size = getimagesize(   public_path()."/img/userImage/high/".$name_image);
-                Image::make($file)->resize($size[0]/3, $size[1]/3)->save($path_medium);
-                Image::make($file)->resize($size[0]/5, $size[1]/5)->save($path_low);
+            $name_image = "userImage_".time().".jpg";
+            $path_high = public_path()."/img/userImage/high/".$name_image;
+            $path_medium = public_path()."/img/userImage/medium/".$name_image;
+            $path_low = public_path()."/img/userImage/low/".$name_image;
 
-                $userImageReceive = UserImage::create([
-                    'path' => $name_image,
-                   // 'resource' => $userImage['resource'],
-                    //falar para eles arrumarem isso
+            Image::make($file)->save($path_high);
+            $size = getimagesize(public_path()."/img/userImage/high/".$name_image);
+            Image::make($file)->resize($size[0]/3, $size[1]/3)->save($path_medium);
+            Image::make($file)->resize($size[0]/5, $size[1]/5)->save($path_low);
 
-                    'orderImage' => $userImage['order'],
-                    'idUser' => $user['idUser'],
-                    'tries' => 0,
-                    'active' => true,
-                    ]);   
-           // }
+            $userImageReceive = UserImage::create([
+                'path' => $name_image,
+                       // 'resource' => $userImage['resource'],
+                'orderImage' => $userImage['orderImage'],
+                'idUser' => $user['idUser'],
+                'tries' => 0,
+                'active' => true,
+                ]);   
+               // }
 
-        return response()->json($userImageReceive);
+            return response()->json($userImageReceive);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -115,8 +156,17 @@ class UserImageController extends Controller
      */
     public function show($id)
     {
-        $all = UserImage::with('user')->findOrFail($id);
-        return response()->json($all);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = UserImage::with('user')->findOrFail($id);
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -128,10 +178,19 @@ class UserImageController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $userImage = UserImage::findOrFail($id);
-        $input = $request->all();
-        $userImage->fill($input)->save();
-        return response()->json($userImage);
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $userImage = UserImage::findOrFail($id);
+            $input = $request->all();
+            $userImage->fill($input)->save();
+            return response()->json($userImage);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -142,7 +201,16 @@ class UserImageController extends Controller
      */
     public function destroy($id)
     {
-        $userImage = UserImage::findOrFail($id);
-        $userImage->delete();
+        $var_token = $_GET['api_token'];
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $userImage = UserImage::findOrFail($id);
+            $userImage->delete();
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 }
